+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Refactor string utility functions to `lib/utils/format.ts`
+- Switched from SQLite to PostgreSQL
 
 ### Deprecated